digital ocean spec group Operation

+++ spec/digitalocean_wrapper/digitalocean_wrapper_spec.rb
@@ -2,7 +2,7 @@ require 'rspec'
 require_relative '../../testing_shared'
 
 describe DigitalOceanWrapper, retry: 1 do
-  let(:digital_ocean) { DigitalOceanWrapper.new }
+  digital_ocean = DigitalOceanWrapper.new
 
   it 'check for correct load access token from file' do
     expect(digital_ocean.client.access_token).not_to be_empty
@@ -47,40 +47,40 @@ describe DigitalOceanWrapper, retry: 1 do
     expect(digital_ocean.current_kernel('testrail')).to eq('Ubuntu 13.10 x64 vmlinuz-3.11.0-12-generic (277)')
   end
 
-  it 'change_kernel' do
-    digital_ocean.restore_image_by_name('nct-at-stable', 'wrapper-test')
-    digital_ocean.wait_until_droplet_have_status('wrapper-test')
-    expect(digital_ocean.current_kernel('wrapper-test')).not_to eq('Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
-    digital_ocean.change_kernel('wrapper-test', 'Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
-    expect(digital_ocean.current_kernel('wrapper-test')).to eq('Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
-    digital_ocean.destroy_droplet_by_name('wrapper-test')
-    expect(digital_ocean.get_droplet_by_name('wrapper-test')).to be_nil
-  end
+  context 'Operation' do
+    before :all do
+      digital_ocean.restore_image_by_name('nct-at-stable', 'wrapper-test')
+      digital_ocean.wait_until_droplet_have_status('wrapper-test')
+    end
 
-  it 'power_off_droplet' do
-    digital_ocean.restore_image_by_name('nct-at-stable', 'wrapper-test')
-    digital_ocean.wait_until_droplet_have_status('wrapper-test')
-    digital_ocean.power_off_droplet('wrapper-test')
-    expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('off')
-    digital_ocean.destroy_droplet_by_name('wrapper-test')
-    expect(digital_ocean.get_droplet_by_name('wrapper-test')).to be_nil
-  end
+    it 'change_kernel' do
+      expect(digital_ocean.current_kernel('wrapper-test')).not_to eq('Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
+      digital_ocean.change_kernel('wrapper-test', 'Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
+      expect(digital_ocean.current_kernel('wrapper-test')).to eq('Ubuntu 14.04 x64 vmlinuz-3.13.0-52-generic')
+    end
 
-  it 'power_on_droplet' do
-    digital_ocean.restore_image_by_name('nct-at-stable', 'wrapper-test')
-    digital_ocean.wait_until_droplet_have_status('wrapper-test')
-    digital_ocean.power_on_droplet('wrapper-test')
-    expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('active')
-    digital_ocean.destroy_droplet_by_name('wrapper-test')
-    expect(digital_ocean.get_droplet_by_name('wrapper-test')).to be_nil
-  end
+    it 'power_off_droplet' do
+      digital_ocean.power_off_droplet('wrapper-test')
+      expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('off')
+    end
 
-  it 'reboot_droplet' do
-    digital_ocean.restore_image_by_name('nct-at-stable', 'wrapper-test')
-    digital_ocean.wait_until_droplet_have_status('wrapper-test')
-    digital_ocean.reboot_droplet('wrapper-test')
-    expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('active')
-    digital_ocean.destroy_droplet_by_name('wrapper-test')
-    expect(digital_ocean.get_droplet_by_name('wrapper-test')).to be_nil
+    it 'power_on_droplet' do
+      digital_ocean.power_on_droplet('wrapper-test')
+      expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('active')
+    end
+
+    it 'reboot_droplet' do
+      digital_ocean.reboot_droplet('wrapper-test')
+      expect(digital_ocean.get_droplet_status_by_name('wrapper-test')).to eq('active')
+    end
+
+    it 'destroy_droplet_by_name' do
+      digital_ocean.destroy_droplet_by_name('wrapper-test')
+      expect(digital_ocean.get_droplet_by_name('wrapper-test')).to be_nil
+    end
+
+    after :all do
+      digital_ocean.destroy_droplet_by_name('wrapper-test') if digital_ocean.get_droplet_by_name('wrapper-test')
+    end
   end
 end
